requirements.txt
opencv-python>=4.7.0
pyserial~=3.5
numpy~=2.0.0
pytesseract~=0.3.10
pillow~=10.3.0